fix(ncp): bump ncloud-sdk-go-v2 to v1.6.29

+++ cloud-control-manager/cloud-driver/drivers/ncp/resources/ImageHandler.go
@@ -176,7 +176,7 @@ func mappingImageInfo(serverImage *vserver.ServerImage) irs.ImageInfo {
 	if len(serverImage.BlockStorageMappingList) > 0 {
 		blockStorageMapping := serverImage.BlockStorageMappingList[0]
 		if blockStorageMapping.BlockStorageSize != nil {
-			blockStorageSize = irs.ConvertByteToGBInt64(*blockStorageMapping.BlockStorageSize)
+			blockStorageSize = irs.ConvertByteToGBInt64(int64(*blockStorageMapping.BlockStorageSize))
 		} else {
 			blockStorageSize = "-1"
 		}
